pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "food-brand-agent"
version = "0.1.0"
description = "Food Brand Intelligence System for India — Multi-agent AI pipeline"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.5.3",
    "playwright>=1.58.0",
    "requests>=2.33.1",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.1.0",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.2",
]

[project.scripts]
food-brand = "agents.main_agent:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "tools*", "skills*"]
